+++ packages/core/icon/glyph/checkbox-indeterminate.js
@@ -19,4 +19,5 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 var CheckboxIndeterminateIcon = function CheckboxIndeterminateIcon(props) {
   return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><g fill-rule="evenodd"><rect fill="currentColor" x="6" y="6" width="12" height="12" rx="2"/><rect fill="inherit" x="8" y="11" width="8" height="2" rx="1"/></g></svg>' }, props));
 };
+CheckboxIndeterminateIcon.displayName = 'CheckboxIndeterminateIcon';
 exports.default = CheckboxIndeterminateIcon;