+++ packages/core/icon/glyph/creditcard-filled.js
@@ -19,4 +19,5 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 var CreditcardFilledIcon = function CreditcardFilledIcon(props) {
   return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><path d="M3 10.988h18V17c0 1.1-.9 2-2 2H5c-1.1 0-2-.9-2-2v-6.012zM6.013 16h2v-2h-2v2zM21 8.993V7c0-1.1-.9-2-2-2H5c-1.1 0-2 .9-2 2v1.993" fill="currentColor" fill-rule="evenodd"/></svg>' }, props));
 };
+CreditcardFilledIcon.displayName = 'CreditcardFilledIcon';
 exports.default = CreditcardFilledIcon;