+++ packages/core/icon/glyph/editor/open.js
@@ -19,4 +19,5 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 var EditorOpenIcon = function EditorOpenIcon(props) {
   return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><path d="M9.873 7.63c-.28 0-.344.159-.147.357l2.436 2.436-4.386 4.386a.509.509 0 0 0 0 .713l.7.7a.495.495 0 0 0 .713.001l4.387-4.386 2.436 2.436c.197.197.357.124.357-.147V8.133a.507.507 0 0 0-.503-.503H9.873z" fill="currentColor" fill-rule="evenodd"/></svg>' }, props));
 };
+EditorOpenIcon.displayName = 'EditorOpenIcon';
 exports.default = EditorOpenIcon;