+++ packages/core/icon/glyph/file-types/24/file-24-archive.js
@@ -19,4 +19,5 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 var FileTypes24File24ArchiveIcon = function FileTypes24File24ArchiveIcon(props) {
   return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><path fill="#5E6C84" fill-rule="evenodd" d="M3 0h18a3 3 0 0 1 3 3v18a3 3 0 0 1-3 3H3a3 3 0 0 1-3-3V3a3 3 0 0 1 3-3zm6 3v3h3V3H9zm3 3v3h3V6h-3zM9 9v3h3V9H9zm3 3v3h3v-3h-3zm-3 3v3h3v-3H9zm3 3v3h3v-3h-3z"/></svg>' }, props));
 };
+FileTypes24File24ArchiveIcon.displayName = 'FileTypes24File24ArchiveIcon';
 exports.default = FileTypes24File24ArchiveIcon;