+++ packages/core/icon/glyph/jira-service-desk.js
@@ -19,4 +19,5 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 var JiraServiceDeskIcon = function JiraServiceDeskIcon(props) {
   return _react2.default.createElement(_index2.default, _extends({ dangerouslySetGlyph: '<svg width="24" height="24" viewBox="0 0 24 24" focusable="false" role="presentation"><defs><linearGradient x1="38.041%" y1="6.637%" x2="59.856%" y2="63.778%" id="a-idPlaceholder"><stop stop-color="inherit" stop-opacity=".4" offset="0%"/><stop stop-color="inherit" offset="100%"/></linearGradient></defs><path d="M13.335 9.98l5.152.02a.803.803 0 0 1 .638 1.285L11.52 21.44a5.402 5.402 0 0 1-1.095-7.563l2.91-3.897z" fill="url(#a-idPlaceholder)"/><path d="M10.422 13.877L5.3 13.845a.803.803 0 0 1-.638-1.285l7.53-10.06a5.402 5.402 0 0 1 1.095 7.563l-2.865 3.814z" fill="currentColor"/></svg>' }, props));
 };
+JiraServiceDeskIcon.displayName = 'JiraServiceDeskIcon';
 exports.default = JiraServiceDeskIcon;