+++ packages/media/media-card/src/root/domain.ts
@@ -9,7 +9,7 @@ export type Provider = MediaItemProvider | UrlPreviewProvider;
 
 export interface FileIdentifier {
   readonly mediaItemType: 'file';
-  readonly id: string;
+  readonly id: string | Promise<string>;
   readonly occurrenceKey?: string;
   readonly collectionName?: string; // files can exist outside of a collection
 }