+++ packages/media/media-picker/example-helpers/upload-preview.tsx
@@ -1,41 +1,27 @@
 import * as React from 'react';
-import {
-  PreviewImage,
-  PreviewImageWrapper,
-  ProgressCircleWrapper,
-} from './styled';
-import { isImagePreview } from '../src/domain/preview';
-import Circle from 'react-circle';
+import { PreviewImageWrapper } from './styled';
 import { PreviewData } from './types';
+import { Card, FileIdentifier } from '@atlaskit/media-card';
+import { createUploadContext } from '@atlaskit/media-test-helpers';
+
+const context = createUploadContext();
 
 export class UploadPreview extends React.Component<PreviewData> {
   render() {
-    const { isProcessed, preview, fileId, uploadingProgress } = this.props;
+    const { upfrontId } = this.props;
 
-    let dimensions;
-    if (isImagePreview(preview)) {
-      dimensions = preview.dimensions;
+    if (!upfrontId) {
+      return <div />;
     }
-    const dimensionsInfo = dimensions ? (
-      <div>
-        Original dimensions ({dimensions.width} x {dimensions.height})
-      </div>
-    ) : null;
+
+    const identifier: FileIdentifier = {
+      id: upfrontId,
+      mediaItemType: 'file',
+    };
 
     return (
       <PreviewImageWrapper>
-        {dimensionsInfo}
-        <PreviewImage fadedOut={!isProcessed} src={preview.src} id={fileId} />
-        {!isProcessed ? (
-          <ProgressCircleWrapper>
-            <Circle
-              textColor="#172B4D"
-              progressColor="#0052CC"
-              bgColor="#DEEBFF"
-              progress={uploadingProgress}
-            />
-          </ProgressCircleWrapper>
-        ) : null}
+        <Card identifier={identifier} context={context} />
       </PreviewImageWrapper>
     );
   }