+++ packages/media/media-picker/src/domain/file.ts
@@ -1,9 +1,11 @@
 export type MediaFile = {
   readonly id: string;
+  readonly upfrontId: Promise<string>;
   readonly name: string;
   readonly size: number;
   readonly creationDate: number;
   readonly type: string;
+  readonly occurrenceKey?: string;
 };
 
 export type PublicMediaFile = MediaFile & {
@@ -11,7 +13,7 @@ export type PublicMediaFile = MediaFile & {
 };
 
 export function copyMediaFileForUpload(
-  { name, size, creationDate, type }: MediaFile,
+  { name, size, creationDate, type, upfrontId, occurrenceKey }: MediaFile,
   uploadId: string,
 ): MediaFile {
   // We dont' use spread here because user upload events are not sanitized
@@ -21,11 +23,21 @@ export function copyMediaFileForUpload(
     size,
     creationDate,
     type,
+    upfrontId,
+    occurrenceKey,
   };
 }
 
 export function copyPublicMediaFileForUpload(
-  { name, size, creationDate, type, publicId }: PublicMediaFile,
+  {
+    name,
+    size,
+    creationDate,
+    type,
+    publicId,
+    upfrontId,
+    occurrenceKey,
+  }: PublicMediaFile,
   uploadId: string,
 ): PublicMediaFile {
   // We dont' use spread here because user upload events are not sanitized
@@ -36,5 +48,7 @@ export function copyPublicMediaFileForUpload(
     creationDate,
     type,
     publicId,
+    upfrontId,
+    occurrenceKey,
   };
 }